Consolidate around `frames` url naming
